Extract ServiceManager internal structure to core package
